refactor(core): split video accessors into HTMLVideoElementLayer
Move poster, picture-in-picture, presentation-mode, and fullscreen
forwarding out of HTMLMediaElementLayer into a new HTMLVideoElementLayer
subclass, and re-root HTMLVideoElementHost on it. Audio hosts no longer
inherit the video surface, aligning the runtime with the Audio/Video
type contracts and trimming the audio-only bundle (simple-hls-audio-only
drops ~268B gzipped, ~18.5%).

Migrate the video-chain layers (hls-js, native-hls, mux, google-cast) to
HTMLVideoElementLayer so they still forward the full video surface, and
narrow the native-hls layers' target overrides to HTMLVideoElement.
