+++ packages/core/src/dom/media/html-media-element-layer.ts
@@ -1,4 +1,3 @@
-import { isFunction } from '@videojs/utils/predicate';
 import { MediaLayer } from '../../core/media/media-layer';
 import type {
   CanPlayTypeResult,
@@ -9,7 +8,6 @@ import type {
   VideoEvents,
 } from '../../core/media/types';
 import { MediaStreamTypes } from '../../core/media/types';
-import type { WebKitPresentationMode, WebKitVideoElement } from '../presentation/types';
 import { EMPTY_REMOTE, EMPTY_TEXT_TRACKS, EMPTY_TIME_RANGES } from './constants';
 
 export abstract class HTMLMediaElementLayer<
@@ -208,61 +206,5 @@ export abstract class HTMLMediaElementLayer<
 
   set disableRemotePlayback(value) {
     if (this.next) this.next.disableRemotePlayback = value;
-  }
-
-  // -- Video --
-
-  get poster() {
-    return this.next?.poster ?? '';
-  }
-
-  set poster(value: string) {
-    if (this.next) this.next.poster = value;
-  }
-
-  get webkitPresentationMode() {
-    return (this.next as WebKitVideoElement | null)?.webkitPresentationMode;
-  }
-
-  get webkitSetPresentationMode(): ((mode: WebKitPresentationMode) => void) | undefined {
-    const fn = (this.next as WebKitVideoElement | null)?.webkitSetPresentationMode;
-    return isFunction(fn) ? fn.bind(this.next) : undefined;
-  }
-
-  get isPictureInPicture(): boolean {
-    const { target } = this;
-    return (
-      (!!target && globalThis.document?.pictureInPictureElement === target) ||
-      this.webkitPresentationMode === 'picture-in-picture'
-    );
-  }
-
-  get isFullscreen(): boolean {
-    const { target } = this;
-    if (!target) return false;
-    if (this.webkitPresentationMode === 'fullscreen') return true;
-    const doc = globalThis.document;
-    if (!doc) return false;
-    return (
-      doc.fullscreenElement === target || ('webkitFullscreenElement' in doc && doc.webkitFullscreenElement === target)
-    );
-  }
-
-  async requestPictureInPicture() {
-    if (!this.next) return Promise.reject();
-    return this.next.requestPictureInPicture();
-  }
-
-  async exitPictureInPicture() {
-    return globalThis.document?.exitPictureInPicture();
-  }
-
-  requestFullscreen() {
-    if (!this.next) return Promise.reject();
-    return this.next.requestFullscreen();
-  }
-
-  exitFullscreen() {
-    return globalThis.document?.exitFullscreen();
   }
 }
